requirements.txt
click==8.0.3
Faker==11.3.0
Flask==2.0.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
PyMySQL==1.0.2
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.29
text-unidecode==1.3
Werkzeug==2.0.2
mysqlclient==2.1.0
pyjwt==2.3.0
bcrypt==3.2.0
python-dotenv==0.19.2
datetime==4.3
